[#20] Make tests use the sample git repo instead
The introduced tests rely on the project clone to have the full history, but the Gitlab CI performs a shallow clone so it won't find old commits.

But we already have a sample full git_repo in `t/data` so we should use it.

I kept the spirit of the original tests implementation.

Signed-off-by: Tomás Cohen Arazi <tomascohen@theke.io>